Extract `CustomLinkText` component
