+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebar.jsx
@@ -32,6 +32,7 @@ import ClickMappings, {
   clickTargetObjectType,
 } from "metabase/dashboard/components/ClickMappings";
 
+import CustomLinkText from "./CustomLinkText";
 import ValuesYouCanReference from "./ValuesYouCanReference";
 import {
   SidebarItemWrapper,
@@ -825,24 +826,5 @@ function QuestionDashboardPicker({ dashcard, clickBehavior, updateSettings }) {
     </div>
   );
 }
-
-const CustomLinkText = ({ clickBehavior, updateSettings }) => {
-  return (
-    <div className="mt2 mb1">
-      <Heading>{t`Customize link text (optional)`}</Heading>
-      <InputBlurChange
-        className="input block full"
-        placeholder={t`E.x. Details for {{Column Name}}`}
-        value={clickBehavior.linkTextTemplate}
-        onBlurChange={e =>
-          updateSettings({
-            ...clickBehavior,
-            linkTextTemplate: e.target.value,
-          })
-        }
-      />
-    </div>
-  );
-};
 
 export default ClickBehaviorSidebar;
